requirements.txt
# Multi-Agent Research & Report Generation System
# Core LangChain / LangGraph packages
langgraph>=0.2.55
langchain-groq>=0.2.4
langchain-community>=0.3.13
langchain-core>=0.3.28

# Tavily web search
tavily-python>=0.5.0

# Streamlit frontend
streamlit>=1.41.1

# Environment variable management
python-dotenv>=1.0.1

# LangSmith tracing (bundled with langchain, listed for clarity)
langsmith>=0.2.3

# PDF export
fpdf2>=2.7.9
